Use mlpack 4.6.3 and its updated mlpack.h
This means we can delete the local copy rcppmlpackexamples.h we had and
update the source files to include <mlpack.h>. We also add a file
rcppmlpackexamples_types.h in order to pass '#define ARMA_USE_CURRENT' to
RcppExports.cpp. Using it via Makevars{,.win} is equivalent, but a little
'louder' on the commandline.